+++ plugin/clickhouse-plugin/src/main/java/org/finos/vuu/plugin/clickhouse/DummyForJavaDoc.java
@@ -1,0 +1,22 @@
+package org.finos.vuu.plugin.clickhouse;
+
+/**
+ * A placeholder class used solely to generate a Javadoc JAR
+ * for publishing this project to Maven Central.
+ *
+ * <p>This class does not provide any functionality and exists
+ * only to satisfy repository requirements.</p>
+ */
+public class DummyForJavaDoc {
+
+    /**
+     * Entry point for the application.
+     * <p>No runtime behaviour is implemented.</p>
+     *
+     * @param args command‑line arguments (ignored)
+     */
+    public static void main(String[] args) {
+        // no-op
+    }
+}
+
